Auditorio de roles y permisos

+++ app/Http/Controllers/OrderController.php
@@ -12,10 +12,13 @@ use App\Mail\CrearPedidoClienteEmail;
 use App\Models\Order;
 use App\Models\OrderProduct;
 use App\Models\ToAsk;
+use App\Models\User;
 use Barryvdh\DomPDF\Facade\Pdf;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Mail;
+
+use Spatie\Permission\PermissionRegistrar;
 
 class OrderController extends \App\Http\Controllers\Controller
 {
@@ -91,28 +94,34 @@ class OrderController extends \App\Http\Controllers\Controller
         }
     }
 
-    public function destroy($id)
+    public function destroy(Request $request)
     {
         DB::beginTransaction();
 
+        app()[PermissionRegistrar::class]->forgetCachedPermissions(); 
+
+        $user = User::find(auth()->user()->id);
+        if (!$user->can('pedido.delete')) {
+            return sendResponse(null, "Acción no autorizada");
+        }
+
         try {
-            $order = Order::findOrFail($id);
-            $entregado = OrderProduct::where('order_id', $id)->where('state_id', 11)->exists();
+            $order = Order::findOrFail($request->id);
+            $entregado = OrderProduct::where('order_id', $request->id)->where('state_id', 11)->exists();
 
             if ($entregado) {
                 throw new \Exception("No se puede eliminar un pedido con un producto entregado");
             }
 
             $order->delete();
-            OrderProduct::where('order_id', $id)->delete();
+            OrderProduct::where('order_id', $request->id)->delete();
 
             DB::commit();
 
-            return sendResponse($id);
+            return sendResponse(new OrderResource($order));
         } catch (\Exception $e) {
             DB::rollBack();
-
-            return sendResponse(null, $e->getMessage(), 300, $id);
+            return sendResponse(null, $e->getMessage(), 300, $request->id);
         }
     }
